数据字典弹框提示，表单验证以及分页显示

+++ we-web-parent/web-linyun-airline/src/main/java/com/linyun/airline/entities/DictRelationEntity.java
@@ -28,6 +28,6 @@ public class DictRelationEntity implements Serializable {
 
 	@Column
 	@Comment("删除标识:(0,表示未删除；1,表示已删除)")
-	private int dr;
+	private long dr;
 
 }
