+++ we-web-parent/web-linyun-airline/src/main/java/com/linyun/airline/forms/DictRelationAddForm.java
@@ -19,5 +19,5 @@ public class DictRelationAddForm extends AddForm implements Serializable {
 	private long targetId;
 
 	/**删除标识,0-未删除，1-已删除*/
-	private int dr;
+	private long dr;
 }
